modularized reactions
